chore: renombrar puente -> paoderado
