+++ packages/core/test/fixtures/extras/apoderado.esjs
@@ -1,0 +1,1 @@
+mut apoderado = crear Apoderado()
