+++ packages/core/test/fixtures/extras/apoderado.js
@@ -1,0 +1,1 @@
+let apoderado = new Proxy()
