--- packages/core/test/fixtures/extras/puente.esjs
@@ -1,1 +1,0 @@
-mut myPuent = crear Puente()
