--- packages/core/test/fixtures/extras/puente.js
@@ -1,1 +1,0 @@
-let myPuent = new Proxy()
